Create OTA.ino
The esp32 creates a server which hosts a webpage, the user can access this webpage by connecting to the esp32 network, and then can access the webpage by typing the IP address "192.168.4.1" or the host name "http://esp32.local()"in their browser . A login page will appear for which the Username and Password is "admin".  Then the user can upload the bin file and click on update to begin the update process.